+++ kotest-extensions/kotest-extensions-testcontainers/src/jvmMain/kotlin/io/kotest/extensions/testcontainers/TestContainerProjectExtension.kt
@@ -2,6 +2,8 @@ package io.kotest.extensions.testcontainers
 
 import io.kotest.core.extensions.MountableExtension
 import io.kotest.core.listeners.AfterProjectListener
+import io.kotest.extensions.testcontainers.options.BasicLogConsumer
+import io.kotest.extensions.testcontainers.options.TestContainerOptions
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.testcontainers.containers.GenericContainer
