+++ kotest-extensions/kotest-extensions-testcontainers/src/jvmMain/kotlin/io/kotest/extensions/testcontainers/options/TestContainerOptions.kt
@@ -1,0 +1,5 @@
+package io.kotest.extensions.testcontainers.options
+
+data class TestContainerOptions(
+   val logs: LogTypes = LogTypes.NONE
+)
